fix(clouds): continuous f64 world-time to kill the wrap seam
The shared cloud world-time was wrapped at rem_euclid(1e6) as f32 to bound
precision. Wind offset and the three weather-drift octaves reduce rate*time
modulo a config-driven period (noise/weather tile), so a fixed wrap is only
seamless when 1e6*rate is a whole multiple of every period -- impossible once
the tiles are tunable. The result was a hard cloud discontinuity every ~11.6
days of world time: the field teleported and coverage flipped at the boundary
(e.g. NYC 2026-05-18 UTC 21:20:00, where day_of_year*86400+utc_seconds hits
12,000,000).

Make CloudWorldTime a continuous f64 (a year of seconds is exact in f64, so no
wrap is needed for precision) and reduce wind + per-octave weather drift from it
in f64. Weather drift rides in as a new per-layer weather_drift vec3 uniform.
time_seconds stays a wrapped f32 for the period-1.0 noise taps (climate, sim,
domain warp), whose rates keep 1e6*rate integral so the wrap is invisible to
them. Only residual seams are the yearly day_of_year rollover and intentional
manual date jumps.

+++ client/ui/src/clouds.rs
@@ -100,7 +100,7 @@ pub(super) fn render_atmosphere_tab(
     }
 }
 
-fn render_overview(ui: &mut egui::Ui, cloud: &mut CloudLayers, world_time: f32) {
+fn render_overview(ui: &mut egui::Ui, cloud: &mut CloudLayers, world_time: f64) {
     ui.label("Quality:");
     ui.horizontal(|ui| {
         for tier in [CloudQuality::Low, CloudQuality::Medium, CloudQuality::High] {
